@@ -1,5 +1,6 @@
 import 'package:auto_size_text/auto_size_text.dart';
 import 'package:basic_calculator/constants.dart';
+import 'package:basic_calculator/logic/calculator_service.dart';
 import 'package:basic_calculator/widgets/calculator_button.dart';
 import 'package:basic_calculator/widgets/calculator_button_text.dart';
 import 'package:flutter/material.dart';
@@ -12,17 +13,26 @@ class Calculator extends StatefulWidget {
 }
 
 class _CalculatorState extends State<Calculator> {
-  int valueA;
-  int valueB;
-  String operator;
-  StringBuffer sb = StringBuffer();
+  CalculatorService _service = CalculatorService();
+
+  callAppendValue(String value) {
+    setState(() {
+      if (value == '0' && _service.stringBuffer.toString() == '0') {
+        return;
+      }
+
+      if (_service.stringBuffer.toString() == '0' && value != '0') {
+        _service.stringBuffer.clear();
+      }
+
+      _service.appendValue(value);
+    });
+  }
 
   @override
   void initState() {
     // TODO: implement initState
     super.initState();
-    sb.write('0');
-    operator = "";
   }
 
   @override
@@ -49,7 +59,7 @@ class _CalculatorState extends State<Calculator> {
                   alignment: Alignment.bottomRight,
                   children: <Widget>[
                     AutoSizeText(
-                      sb.toString(),
+                      _service.stringBuffer.toString(),
                       style: Theme.of(context).textTheme.display2,
                       maxLines: 1,
                     )
@@ -73,17 +83,46 @@ class _CalculatorState extends State<Calculator> {
                             text: 'C',
                             color: kPrimarySwatchColor,
                           ),
+                          onPressed: () {
+                            setState(() {
+                              _service.stringBuffer.clear();
+                              _service.stringBuffer.write('0');
+                            });
+                          },
                         ),
                         CalculatorButton(
-                            hairlineColor: kHairlineColor,
-                            primarySwatchColor: Colors.white,
-                            flex: 1,
-                            child: Icon(Icons.backspace)),
+                          hairlineColor: kHairlineColor,
+                          primarySwatchColor: Colors.white,
+                          flex: 1,
+                          child: Icon(Icons.backspace),
+                          onPressed: () {
+                            setState(() {
+                              var values = _service.stringBuffer
+                                  .toString()
+                                  .split('')
+                                  .toList();
+
+                              if (values.length > 1) {
+                                values.removeLast();
+                              } else {
+                                values = new List();
+                                values.add('0');
+                              }
+
+                              _service.stringBuffer.clear();
+                              _service.stringBuffer.write(values.join());
+                            });
+                          },
+                        ),
                         CalculatorButton(
-                            hairlineColor: kHairlineColor,
-                            primarySwatchColor: Colors.white,
-                            flex: 1,
-                            child: CalculatorButtonText(text: '/'))
+                          hairlineColor: kHairlineColor,
+                          primarySwatchColor: Colors.white,
+                          flex: 1,
+                          child: CalculatorButtonText(text: '/'),
+                          onPressed: () {
+                            callAppendValue('/');
+                          },
+                        )
                       ],
                     )),
                     Expanded(
@@ -91,25 +130,41 @@ class _CalculatorState extends State<Calculator> {
                       crossAxisAlignment: CrossAxisAlignment.stretch,
                       children: <Widget>[
                         CalculatorButton(
-                            hairlineColor: kHairlineColor,
-                            primarySwatchColor: Colors.white,
-                            flex: 1,
-                            child: CalculatorButtonText(text: '7')),
+                          hairlineColor: kHairlineColor,
+                          primarySwatchColor: Colors.white,
+                          flex: 1,
+                          child: CalculatorButtonText(text: '7'),
+                          onPressed: () {
+                            callAppendValue('7');
+                          },
+                        ),
                         CalculatorButton(
-                            hairlineColor: kHairlineColor,
-                            primarySwatchColor: Colors.white,
-                            flex: 1,
-                            child: CalculatorButtonText(text: '8')),
+                          hairlineColor: kHairlineColor,
+                          primarySwatchColor: Colors.white,
+                          flex: 1,
+                          child: CalculatorButtonText(text: '8'),
+                          onPressed: () {
+                            callAppendValue('8');
+                          },
+                        ),
                         CalculatorButton(
-                            hairlineColor: kHairlineColor,
-                            primarySwatchColor: Colors.white,
-                            flex: 1,
-                            child: CalculatorButtonText(text: '9')),
+                          hairlineColor: kHairlineColor,
+                          primarySwatchColor: Colors.white,
+                          flex: 1,
+                          child: CalculatorButtonText(text: '9'),
+                          onPressed: () {
+                            callAppendValue('9');
+                          },
+                        ),
                         CalculatorButton(
-                            hairlineColor: kHairlineColor,
-                            primarySwatchColor: Colors.white,
-                            flex: 1,
-                            child: CalculatorButtonText(text: 'X')),
+                          hairlineColor: kHairlineColor,
+                          primarySwatchColor: Colors.white,
+                          flex: 1,
+                          child: CalculatorButtonText(text: 'x'),
+                          onPressed: () {
+                            callAppendValue('x');
+                          },
+                        ),
                       ],
                     )),
                     Expanded(
@@ -117,25 +172,41 @@ class _CalculatorState extends State<Calculator> {
                       crossAxisAlignment: CrossAxisAlignment.stretch,
                       children: <Widget>[
                         CalculatorButton(
-                            hairlineColor: kHairlineColor,
-                            primarySwatchColor: Colors.white,
-                            flex: 1,
-                            child: CalculatorButtonText(text: '4')),
+                          hairlineColor: kHairlineColor,
+                          primarySwatchColor: Colors.white,
+                          flex: 1,
+                          child: CalculatorButtonText(text: '4'),
+                          onPressed: () {
+                            callAppendValue('4');
+                          },
+                        ),
                         CalculatorButton(
-                            hairlineColor: kHairlineColor,
-                            primarySwatchColor: Colors.white,
-                            flex: 1,
-                            child: CalculatorButtonText(text: '5')),
+                          hairlineColor: kHairlineColor,
+                          primarySwatchColor: Colors.white,
+                          flex: 1,
+                          child: CalculatorButtonText(text: '5'),
+                          onPressed: () {
+                            callAppendValue('5');
+                          },
+                        ),
                         CalculatorButton(
-                            hairlineColor: kHairlineColor,
-                            primarySwatchColor: Colors.white,
-                            flex: 1,
-                            child: CalculatorButtonText(text: '6')),
+                          hairlineColor: kHairlineColor,
+                          primarySwatchColor: Colors.white,
+                          flex: 1,
+                          child: CalculatorButtonText(text: '6'),
+                          onPressed: () {
+                            callAppendValue('6');
+                          },
+                        ),
                         CalculatorButton(
-                            hairlineColor: kHairlineColor,
-                            primarySwatchColor: Colors.white,
-                            flex: 1,
-                            child: CalculatorButtonText(text: '-')),
+                          hairlineColor: kHairlineColor,
+                          primarySwatchColor: Colors.white,
+                          flex: 1,
+                          child: CalculatorButtonText(text: '-'),
+                          onPressed: () {
+                            callAppendValue('-');
+                          },
+                        ),
                       ],
                     )),
                     Expanded(
@@ -143,25 +214,41 @@ class _CalculatorState extends State<Calculator> {
                       crossAxisAlignment: CrossAxisAlignment.stretch,
                       children: <Widget>[
                         CalculatorButton(
-                            hairlineColor: kHairlineColor,
-                            primarySwatchColor: Colors.white,
-                            flex: 1,
-                            child: CalculatorButtonText(text: '1')),
+                          hairlineColor: kHairlineColor,
+                          primarySwatchColor: Colors.white,
+                          flex: 1,
+                          child: CalculatorButtonText(text: '1'),
+                          onPressed: () {
+                            callAppendValue('1');
+                          },
+                        ),
                         CalculatorButton(
-                            hairlineColor: kHairlineColor,
-                            primarySwatchColor: Colors.white,
-                            flex: 1,
-                            child: CalculatorButtonText(text: '2')),
+                          hairlineColor: kHairlineColor,
+                          primarySwatchColor: Colors.white,
+                          flex: 1,
+                          child: CalculatorButtonText(text: '2'),
+                          onPressed: () {
+                            callAppendValue('2');
+                          },
+                        ),
                         CalculatorButton(
-                            hairlineColor: kHairlineColor,
-                            primarySwatchColor: Colors.white,
-                            flex: 1,
-                            child: CalculatorButtonText(text: '3')),
+                          hairlineColor: kHairlineColor,
+                          primarySwatchColor: Colors.white,
+                          flex: 1,
+                          child: CalculatorButtonText(text: '3'),
+                          onPressed: () {
+                            callAppendValue('3');
+                          },
+                        ),
                         CalculatorButton(
-                            hairlineColor: kHairlineColor,
-                            primarySwatchColor: Colors.white,
-                            flex: 1,
-                            child: CalculatorButtonText(text: '+')),
+                          hairlineColor: kHairlineColor,
+                          primarySwatchColor: Colors.white,
+                          flex: 1,
+                          child: CalculatorButtonText(text: '+'),
+                          onPressed: () {
+                            callAppendValue('+');
+                          },
+                        ),
                       ],
                     )),
                     Expanded(
@@ -169,15 +256,23 @@ class _CalculatorState extends State<Calculator> {
                       crossAxisAlignment: CrossAxisAlignment.stretch,
                       children: <Widget>[
                         CalculatorButton(
-                            hairlineColor: kHairlineColor,
-                            primarySwatchColor: Colors.white,
-                            flex: 3,
-                            child: CalculatorButtonText(text: '0')),
+                          hairlineColor: kHairlineColor,
+                          primarySwatchColor: Colors.white,
+                          flex: 3,
+                          child: CalculatorButtonText(text: '0'),
+                          onPressed: () {
+                            callAppendValue('0');
+                          },
+                        ),
                         CalculatorButton(
-                            hairlineColor: kHairlineColor,
-                            primarySwatchColor: Colors.white,
-                            flex: 1,
-                            child: CalculatorButtonText(text: '='))
+                          hairlineColor: kHairlineColor,
+                          primarySwatchColor: Colors.white,
+                          flex: 1,
+                          child: CalculatorButtonText(text: '='),
+                          onPressed: () {
+                            callAppendValue('=');
+                          },
+                        )
                       ],
                     ))
                   ],
@@ -186,47 +281,5 @@ class _CalculatorState extends State<Calculator> {
         ),
       ),
     );
-  }
-
-  void appendValue(String string) {
-    setState(() {
-      bool isDoCalculate = false;
-      if (string == '=') {
-        isDoCalculate = true;
-      } else if (string == '/' ||
-          string == 'X' ||
-          string == '-' ||
-          string == '+') {
-        operator = string;
-      }
-
-      if (!isDoCalculate) {
-        sb.write(string);
-      } else {
-        List<String> values = sb.toString().split(operator);
-        if (values.length == 2 &&
-            values[0].isNotEmpty &&
-            values[1].isNotEmpty) {
-          valueA = int.parse(values[0]);
-          valueB = int.parse(values[1]);
-          sb.clear();
-          int total = 0;
-          switch (operator) {
-            case '/':
-              total = valueA ~/ valueB;
-              break;
-            case 'X':
-              total = valueA * valueB;
-              break;
-            case '-':
-              total = valueA - valueB;
-              break;
-            case '+':
-              total = valueA + valueB;
-          }
-          sb.write(total);
-        }
-      }
-    });
   }
 }
